Add block textures and very simple world generation

+++ Craft/Chunk.h
@@ -15,13 +15,13 @@ constexpr int CHUNKS_Z = 4;
 
 struct Chunk
 {
-	uint8_t blocks[BLOCKS_X][BLOCKS_Y][BLOCKS_Z];
+	int blocks[BLOCKS_X][BLOCKS_Y][BLOCKS_Z];
 	GLuint vbo;
 	int n_vertices;
 	bool changed;
 
-	uint8_t get(int x, int y, int z);
-	void set(int x, int y, int z, uint8_t block_type);
+	int get(int x, int y, int z);
+	void set(int x, int y, int z, int block_type);
 
 	void update();
 	void render(Display* display);
@@ -34,8 +34,8 @@ struct SuperChunk
 {
 	Chunk* c[CHUNKS_X][CHUNKS_Y][CHUNKS_Z];
 
-	uint8_t get(int x, int y, int z);
-	void set(int x, int y, int z, uint8_t block_type);
+	int get(int x, int y, int z);
+	void set(int x, int y, int z, int block_type);
 
 	void render(Display* display);
